Add API for doing a local broadcast to PEs on a node
This respects [nokeep] and therefore would allow a receiver on a node to
forward a message to all PEs on a node, without copying the message, and
automatically deallocating the message after all PEs have received it (similar
to [nokeep] array broadcasts within a PE). With more work this could be more
integrated into all broadcasts/multicasts to save copying costs for large
readonly broadcasts.

Change-Id: I04dd6867245cd6fd8cd59b6125f9e806e5c27d52